Improve execute (#19)
Improves execute by fetching logs on an interval and printing them to stderr,
It does this by implementing a logs `Watcher`. 

The watcher maintains a background go-routine where it fetches the task
logs, status and sends them as RunState on a channel that the execute
command can consume and use.