Prompt users for a slug when creating from a sample (#46)
This will now prompt users for a slug when creating a task from a sample. In a future PR, I'll add logic for validating that a slug is unique, and at that point I imagine we can make this prompt conditional on the slug not being unique already.

```
❯ lairlocal tasks init
Airplane is a development platform for engineers building internal tools.

This command will configure a task definition which Airplane uses to deploy your task.

? How do you want to get started? Create from an Airplane-provided sample
? Pick a runtime: Node.js
? Pick a sample: Hello World
? Pick a unique identifier (slug) for this task: hello_world

An Airplane task definition for 'Hello World' has been created!

To deploy it to Airplane, run:
  airplane tasks deploy -f hello-world-javascript/airplane.yml
```

For this PR, I brought back some YAML parsing logic that I had used (and removed) in a previous PR. We can use this to update fields in a YAML document without losing comments/formatting/etc. that is lost when (un)marshaling.