Add support for uploading to the remote builder (#57)
This PR adds a new `--builder` flag to the `deploy` command which allows you to toggle between the local builder (via the local Docker daemon) or the remote builder (via the Airplane hosted Docker builder). For now, we default to `--builder=local` since the remote builder isn't live yet.

As for remote building functionality, this PR adds logic for archiving and uploading source code into GCS where it'll be available to our hosted builder. Once it does that, it temporarily errors until the remote builder is ready.

I've left a number of TODO comments that will be addressed in future passes, but let me know if any aren't clear.