Lock base image versions to digest SHAs (#72)
We now have a `versions.json` file that is embedded in our CLI and contains a lookup table for identifying the image, tag, and digest to use for any given builder and user-supplied builder version (currently that's just Node, the rest are hardcoded since the major versions are stable enough that there is no relevant choice for customers to make).

We store this information in JSON in a centralized file (rather than split across our builders) since this gives us a format we can easily automate down the road to make upgrading versions easier.

While here, I also made the decision to a) standardize on Debian (buster) b) standardize on specifying minor versions. The former impacted Go and Deno, which previously were on Alpine. The latter is a no-op, considering we're locking on the digest, but it helps clarify what underlying version the digest maps to.