Add support for `.dockerignore` files when constructing uploads (#80)
If you have a `.dockerignore` in the root of your task directory, we will now respect it when creating an upload of your task directory.

Note this change is only relevant to how we upload your task directory, not the docker build process itself. When performing a local build, the docker CLI will use the `.dockerignore` like normal and when performing a remote build, the files ignored by the `.dockerignore` simply will not exist since they were not included in the upload.

This PR also adds defaults `.dockerignore` files for each builder we support so that customers only need to add their own `.dockerignore` if they want custom behavior.

--

This PR uses a fork of `mholt/archiver` to add support for filtering archives: https://github.com/airplanedev/archiver/pull/1

We might want to upstream this at some point, but we would likely need to expand support across all other archive formats first.