Fix Go builder error with multi-file main packages (and add more debug info) (#84)
I ran into an issue where the Go builder was failing to compile our Run SQL example. After doing some digging, I found a fix -- the Go entrypoint should be the main package unless there's only a single file in the main pkg.

While debugging, I ended up making some cleanups that I've included here:

## Docker Logs

Previously, we only showed these logs if you ran with the `--debug` flag. We now parse the JSON-based Docker events from both building and pushing and render them to stderr.

## Failing on docker errors

Now that we parse Docker logs, we can detect if the build or push failed since this information is only conveyed in the logs (rather than via an `error` from the Docker client). For example, previously if a builder failed at any point in executing a Dockerfile, you got a confusing error -- that the tag could not be found.

![image](https://user-images.githubusercontent.com/2907397/115082359-cd06e000-9ed3-11eb-9fc6-fa5e37a3ef44.png)


## Print Dockerfile when `--debug`

We now log the generated Dockerfile when run with `--debug`.

## Multi-line debug prefix

Since Dockerfiles span multiple lines, the `[debug]` prefix was only rendering on the first line. To make this look nice, I've updated the `logger.Debug` function to split the debug message by newlines in order to insert the `[debug]` prefix on every line.