Change YAML file format (#105)
Before:
```
builder: python
builderConfig:
  entrypoint: main.py
```

After:
```
python:
  entrypoint: main.py
```

This PR adds logic for understanding multiple yaml file versions; the latest version will be tried first, & then previous versions. Only errors from the newest version will be shown to the user. This doesn't change the output of `tasks get`; communications with the API server are not changing, either. When initializing from a sample, if the sample's task definition is an older version, the new task definition will be upgraded in the newest version.

Bonus: `ResourceLimits` -> `ResourceRequests`.