Bring back `--file/-f` flag to `deploy/execute` as a deprecated flag (#163)
This PR brings back the `--file/-f` flag to the `deploy` and the `execute` commands. This was previously going to be a breaking change between `0.0.x` and `0.1.x`, but we want to reduce the incompatibility between these releases.

When used, it will log a warning message:

![image](https://user-images.githubusercontent.com/2907397/121256473-98cbf200-c87a-11eb-914b-07366632e962.png)

While here, I also fixed an issue with `execute`-ing slugs that match a local file. I also added the short/long command descriptions to our custom command help text.