Clarify `do not edit` nature of linking comment (#161)
This PR updates the comment we use to link files with Airplane tasks. It now uses the copy from: https://github.com/airplanedev/cli/pull/148#discussion_r647680710

While here, I also extracted the slug detection logic from the JS runtime and tweaked it to handle linking comments that are not on the first line of the script file. This resolves: https://linear.app/airplane/issue/AIR-1228/detect-missing-slug-comment

---

I didn't bother to retain backwards compatibility with the previous version of this comment, since we haven't released it yet. For internal scripts, you'll want to update the comment:

```js
// Linked to Airplane task, do not edit:
// https://app.airplane.so:5000/t/outputs

// ->

// Linked to https://app.airplane.so:5000/t/outputs [do not edit this line]
```