Fix issue with single quotes in Node shim (#201)
This PR fixes a bug introduced in https://github.com/airplanedev/cli/pull/199, which has not yet been released.

That PR added comments to the shim file that contained single quotes, which broke the Dockerfile we generated for Node tests. That PR had just been tested with `airplane dev`, which does not perform a Docker build.