Detect expired CLI tokens (#230)
This PR detects if a user's token is expired and sends them through the login flow again. This will happen every 3 months. We should eventually support auto-refreshing tokens.

This PR also adds a logger to `retryablehttp` that prints out HTTP logs when `--debug` is enabled.

This PR also tweaks `airplane login` to re-login even if currently logged in so that you don't need to `airplane logout` first.