Add `--since/--until` filtering to `runs list` (#235)
This PR updates `runs list` s.t. you can optionally fetch all runs. It does this by introducing a `--since`, `--until`, and `--limit` flag. `--limit` defaults to `100`, but can be overridden to `0` (or any negative number) to fetch all runs. This can be optionally combined with `--since` and `--until` to filter by a time range.

For example, you could list all runs with:

```
airplane runs list --until $(date +%Y-%m-%dT%H:%M:%S%z) --limit=0
```

To support time-based flags, which cobra does not support out-of-the-box (!), I've added a `utils.TimeValue` which is a `pflag.Value`. This means you can use it with `cmd.Flags().Var(&timeValue, ...)`. It tries to be generous about parsing times and provide useful error messages.

The auto-pagination is a bit complicated. Would be a good use-case for generics!