Loosen default airplane ignores (#239)
This was triggered by bin/ being ignored for rails, which causes `rails
runner` and other commands to not work. (Rails has some magic where it
looks for and runs the local bin/).

In hindsight, I think ignoring some of these (e.g. *.env) was too
aggressive, and we should only ignore the ones that are known to be not
necessary (or dangerous even) to include in builds.