Mark pg-native as external as a temporary workaround (#274)
Until we figure out a more general solution to this, this tells esbuild
to NOT compile in pg-native.

The pg package does `require("pg-native)`, but it marks pg-native as an
optional peer dependency. If it can't require it, it silently skips.

However, esbuild tries to bundle it in, and thus it hits errors. It
seems like folks have solved this in other bundlers like webpack by
ignoring the module, shimming it, etc.

The temporary solution is to mark it as external, so we don't attempt to
build it into the bundle.