Mark all node dependencies as external for esbuild (#275)
Read package.json and return all dependencies and devDependencies as
external for esbuild.

This fixes the issue where e.g. pg requires pg-native but gracefully
handles the module not being installed. esbuild will error on this when
trying to bundle.

We don't really need esbuild to bundle _anything_, so we just mark all
dependencies as external.