Normalize entrypoint to match case on filesystem (#279)
First does a refactor to clean up some Definition code. Removes an extraneous
`filepath.Clean` call, which isn't needed because all of the calls around it
clean the path already.

Adds a fix to find the proper casing of entrypoint by listing the directory and
comparing (case-insensitive) to get the proper filesystem casing.

This fixes a bug where a user can `airplane deploy somescript.js`, the filename
is called `someScript.js`, and we later error trying to open `somescript.js` on
a case-sensitive filesystem (e.g. ext4 on linux).