Only deploy tasks that have changed files (#337)
Add a flag that accepts a csv of changed files. If the task (looking from the root dir) does not contain any of these changed files, do not deploy it.

We can be 100% confident that if a task has no mutated files in its root directory, it does not have to be redeployed.