Handle deploying from a task definition file (#340)
I punted on handling explicit permissions and new tasks, will address in later PRs. There's some thought towards handling multiple task definition files, but it's not fully fleshed out. This patch works for both local + remote builds.

I don't know how well this works for complicated Root determination, & what exactly should happen if the task defn file is not in the working directory & the entrypoint is not in the same directory as the task defn file. I suspect this doesn't handle that well & I'm not sure I understand the concept of the Root well enough to figure out exactly what should happen here. This works for the case where the working directory is where the task defn file + entrypoint are. Filed a ticket for examining Root complications in more detail.

Resolves AIR-2462, AIR-2546, AIR-2463.